# backend/src/main/java/com/eaglebank/backend/dto/BadRequestErrorResponse.java
package com.eaglebank.backend.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class BadRequestErrorResponse {
    private String message;
    private List<Detail> details;

    @Data
    @NoArgsConstructor
    @AllArgsConstructor
    @Builder
    public static class Detail {
        private String field;
        private String message;
        private String type;
    }
}