# backend/src/main/java/com/eaglebank/backend/dto/ErrorResponse.java
package com.eaglebank.backend.dto;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class ErrorResponse
{
    private String message;
}