# backend/src/main/java/com/eaglebank/backend/dto/ListBankAccountsResponse.java
package com.eaglebank.backend.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class ListBankAccountsResponse {
    private List<BankAccountResponse> accounts;
}